requirements.txt
requests
python-dotenv
tzlocal==2.1
flask_apscheduler
redis
kafka-python-ng
watchdog
flask

